+++ examples/example-rest/src/main/scala/com/webtrends/service/rest/ReadJSON.scala
@@ -53,7 +53,7 @@ class ReadJSON extends Command
     val p = Promise[CommandResponse[T]]
     bean match {
       case Some(b) =>
-        getComponent("wookie-cache-memcache") onComplete {
+        getComponent("wookiee-cache-memcache") onComplete {
           case Success(actor) =>
             val personName = b("name").asInstanceOf[String]
             Person(personName).readFromCache(actor) onComplete {