+++ examples/example-rest/src/main/scala/com/webtrends/service/sockorest/SRead.scala
@@ -61,7 +61,7 @@ class SRead extends Command
     val p = Promise[CommandResponse[T]]
     bean match {
       case Some(b) =>
-        getComponent("wookie-cache-memcache") onComplete {
+        getComponent("wookiee-cache-memcache") onComplete {
           case Success(actor) =>
             val personName = b("name").asInstanceOf[String]
             Person(personName).readFromCache(actor) onComplete {